% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dietary}
\alias{Dietary}
\title{A diet-microbiome data}
\format{
A list with OTU table and other covariates
\describe{
  \item{OTU}{OTU table}
  \item{COV}{Major covariates extracted from original data}
  \item{M}{Sequencing depth}
}
}
\source{
\url{https://github.com/knights-lab/dietstudy_analyses}
}
\usage{
Dietary
}
\description{
A diet-microbiome data with shotgun metagenomic sequencing results of fecal samples and daily dietary records of 34 subjects in 17 consecutive days.
See original data from article "Daily Sampling Reveals Personalized Diet-Microbiome Associations in Humans" (Johnson et al. 2019).
}
\keyword{datasets}
