% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{pval_getpval}
\alias{pval_getpval}
\title{pscl function to get p-value from pscl result}
\usage{
pval_getpval(object)
}
\arguments{
\item{object}{pscl result}
}
\value{
pvalue
}
\description{
pscl function to get p-value from pscl result
}
