% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{pwald}
\alias{pwald}
\title{Get Wald test p-value}
\usage{
pwald(res, test_index)
}
\arguments{
\item{res}{Output from ZIPG_optim_EM}

\item{test_index}{Index of which parameters equal to zero under H0.}
}
\value{
A list of Wald test result
}
\description{
Get Wald test p-value
}
