% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_init_pscl}
\alias{ZIPG_init_pscl}
\title{Initialize ZIPG from pscl estimation}
\usage{
ZIPG_init_pscl(W, M, X, X_star, A = 1, d, d_star, return_model = T)
}
\arguments{
\item{return_model}{}
}
\description{
Initialize ZIPG from pscl estimation
}
