% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_logli_EM}
\alias{ZIPG_logli_EM}
\title{Log-likelihood for ZIPG given latent variable}
\usage{
ZIPG_logli_EM(W, M, X, X_star, A, d, d_star, parms, zp)
}
\arguments{
\item{zp}{}
}
\description{
Log-likelihood for ZIPG given latent variable
}
