% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_main_EM}
\alias{ZIPG_main_EM}
\title{Run EM algorithm to fit ZIPG model.}
\usage{
ZIPG_main_EM(
  W,
  M,
  X,
  X_star,
  optim_method = "BFGS",
  init = NULL,
  return_model = T,
  ref_init = NULL
)
}
\arguments{
\item{W}{Count data}

\item{M}{Sequencing depth, ZIPG use log(M) as offset by default}

\item{X}{Formula of covariates of differential abundance}

\item{X_star}{Formula of covariates of differential varibility}

\item{optim_method}{Default : 'BFGS'}

\item{init}{initialization parameters}

\item{return_model}{whether return full complete imfomation for fitted model}

\item{ref_init}{Reference}
}
\value{
A list of EM optimize result
}
\description{
Run EM algorithm to fit ZIPG model.
}
