% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_optim_EM}
\alias{ZIPG_optim_EM}
\title{ZIPG EM algorithm to optimize log-likelihood}
\usage{
ZIPG_optim_EM(
  W,
  M,
  X,
  X_star,
  A = 1,
  d,
  d_star,
  parms,
  optim_method = "BFGS",
  fix_index = NULL,
  tol = 1e-05,
  maxit = 200
)
}
\description{
ZIPG EM algorithm to optimize log-likelihood
}
