% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_summary}
\alias{ZIPG_summary}
\title{Summary for ZIPG_main() result.}
\usage{
ZIPG_summary(ZIPG_res, type = "Wald")
}
\arguments{
\item{ZIPG_res}{Result from ZIPG_main()}

\item{type}{Type of hypothesis testing method, 'Wald','bWald' or 'pbWald'.}
}
\value{
pvalue
}
\description{
Summary for ZIPG_main() result.
}
\examples{
data(Dietary)
dat = Dietary
ZIPG_res <- ZIPG_main(data = dat$COV,
X = ~ALC01+nutrPC1+nutrPC2, X_star = ~ ALC01,
W = dat$OTU[,100], M = dat$M )
ZIPG_summary(ZIPG_res)
}
