% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_logli}
\alias{ZIPG_logli}
\title{Log-likelihood for ZIPG}
\usage{
ZIPG_logli(W, M, X, X_star, A = 1, d, d_star, parms, sep = FALSE)
}
\arguments{
\item{sep}{}
}
\description{
Log-likelihood for ZIPG
}
