% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_main_EM}
\alias{ZIPG_main_EM}
\title{Run EM algorithm to fit ZIPG model.}
\usage{
ZIPG_main_EM(
  W,
  M,
  X,
  X_star,
  optim_method = "BFGS",
  init = NULL,
  return_model = T,
  ref_init = NULL
)
}
\arguments{
\item{ref_init}{}
}
\description{
Run EM algorithm to fit ZIPG model.
}
