% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_main.R
\name{ZIPG_optim_EM}
\alias{ZIPG_optim_EM}
\title{ZIPG EM algorithm to optimize log-likelihood}
\usage{
ZIPG_optim_EM(
  W,
  M,
  X,
  X_star,
  A = 1,
  d,
  d_star,
  parms,
  optim_method = "BFGS",
  fix_index = NULL,
  tol = 1e-05,
  maxit = 200
)
}
\arguments{
\item{W}{Count data}

\item{M}{Sequencing depth, ZIPG use log(M) as offset by default}

\item{X}{Formula of covariates of differential abundance}

\item{X_star}{Formula of covariates of differential varibility}

\item{A}{(No use, set A=1)}

\item{d}{The number of covariates of differential abundance}

\item{d_star}{The number of covariates of differential varibility}

\item{parms}{initialization parameters}

\item{optim_method}{Default : 'BFGS'}

\item{fix_index}{Index of parameters that are fixed in optimization}

\item{tol}{convergence criterion}

\item{maxit}{maximum number of iteration}
}
\description{
ZIPG EM algorithm to optimize log-likelihood
}
